algorithm/hybrid_search_khoa/requirements.txt
streamlit>=1.32.0
sentence-transformers>=2.5.1
faiss-cpu>=1.9.0
rank-bm25==0.2.2
numpy